Feature/sample UI wallet (#30)
* Added demo scene with barebones connect UI page. Created implementation for a basic scale tweening animation.

* Added logic to switch UI pages, supports 'back' (needs to be mapped to the back key on various platforms, but for now, we have a UI button). Added a very minimal login screen for testing this out.

* Added a UIPage class that contains the base logic for a UI screen/page (e.g. open/close methods) and made ConnectPage and LoginPage inherrit from it

* Refactor SequenceUI so that it no longer uses UIStates and instead calls the open/close methods on the UIPages directly.

* Add basic UI. For some reason this wasn't saved properly and wasn't included in my previous commit and I had to rebuild it... *sigh*

* Added a panel behind the UI

* Moved UI scripts into a folder

* Added the ability to choose from a set of tweening presets (implementing the ITween interface) to animate the various UI pages. The main UI panel now scrolls in from the bottom

* Wait for panel to animate onto the screen before animating in the first UI page

* Basic login page

* Made the login page the first page and hooked up buttons on login page such that they debug.log what they are supposed to do when clicked

* Added a color scheme manager that allows you to automatically reskin the UI by creating a scriptable object via 'Assets>Create>Sequence>Color Scheme', setting it as the color scheme in the ColorSchemeManager monobehaviour/gameobject, and clicking Apply in the inspector

* Added dark colour scheme

* Added login success page

* Added login handler to the login page - this is an ILogin and is designed to handle the login request. Currently, mocking the login. Upon success or failed login, an ILogin will emit an appropriate event. SequenceUI is a subscriber to this event.

* Move to login success page upon login success. Fixed UI transition animations (didn't have animations out)

* Added an IValidator interface that is used to validate inputs. For now, there is only email validation - I have both a real (with tests) and mocked implementation of IValidator

* Moved SequenceIndexer into its own assembly - referencing SequenceEthereum. Created assembly for SequenceExamples - referencing all the other Sequence assemblies (minus test assemblies) in the project. Finally, create a SequenceExamplesTests assembly for any tests related to the SequenceExamples and renamed our other test assembly from Tests to SequenceTests for clarity.

* Created a new directory (Authentication) and assembly (SequenceAuthentication) under SequenceSDK and moved the authentication-related logic out of SequenceExamples and into Authentication.

* Created base MFA page

* Updated readme

* Display MFA page and add mocked functionality to it

* Fixed runtime error and erase 2fa code whenever revisiting the page

* Only display back button when usable

* Don't overwrite text that is transparrent with different colors

* Moved SequenceUI monobehaviour off of the Panel and onto the canvas, made it stop being a UIPage. Instead, created a UIPanel which inherrits from UIPage - this is used to create the LoginPanel and will be used for the WalletPanel as well

* Setup for UI integration tests

* Setup for multiple UI tests

* Added additional integration tests for sample UI login flow. Run all UI tests in one test so we don't have to setup and teardown the scene each time - makes test suite faster.

* Updated Makefile to run both edit mode and play mode tests when running 'make test'

* Added scroll view, creating NFT prefab (to be instantiated), created generic object pool

* Added 'powered by sequence' to the bottom of the grid layout

* UIPanels hold their initial page

* Added support for newly supported indexer chains

* Added IIndexer and implementation (ChainIndexer) - this is basically just Indexer.cs with the static methods extracted, no longer requiring chain id as a parameter, instead chain Id is provided in the constructor.

* Fetch nft content when opening WalletPage - for now this is mocked

* Switch to event-driven implementation for fetching as opposed to coroutine

* Instantiate images in wallet view

* Don't make ObjectPool a monobehaviour

* Fix scaling issues

* Fix animation

* Subscribe to scroll event

* Update scroll view size based on number of fetched objects

* Allow object pool to grow. Previously wanted to recycle the elements in the pool; however, since the layout of the objects is managed by the grid layout group, anytime one of the elements is removed, the rest are shifted up and are also removed. Instead, we simply allow the object pool to grow as needed - this should still perform well on initial load, but has the potential to create slowdowns if the user has a bunch of NFTs to display.

* Cleanup object pool when closing wallet page

* Navigate to wallet panel and wallet page

* Setup UI test harness so that we can run separate tests for the login flow and the wallet flow. Running these separately (as opposed to all in one test) feels like a happy middleground to me in terms of test runtime vs clarity (to pinpoint the issue) when tests fail.

* Moved opening of next panel into a separate function that has a separate on click listener for the buttons. The functionality is the same as before, except, now it is easier to customize (developers can simply disable the Open{PanelName} on click listener and replace with some other panel to open next)

* Test that we fetch enough items (random between 0 and 1000) and test that the branding stays at the bottom of the page

* Fix ColorSchemeManager such that all panels have their colors modified when applying a new color scheme

* Populate tokens in wallet - using a mocked fetcher

* Fix token name formatting

* Apply color scheme to token ui elements as well

* Refresh token currency values every x seconds - using mock to generate prices

* Fixed broken UI tests

* Fixed ColorSchemeManager such that the changes applied to the prefab is saved.

* Test that percentchange text is displayed appropriately

* Added basic navigation bar to wallet panel. Tested that we can close and re-open the wallet panel. Fixed some issues with the implementation

+++ Assets/SequenceExamples/Scripts/SequenceDemo.cs
@@ -25,7 +25,7 @@ public class SequenceDemo : MonoBehaviour
                     true,
                     new Page { page = pageNumber }
                 );
-                var balances = await Indexer.GetTokenBalances(Chain.Polygon, tokenBalancesArgs);
+                var balances = await Indexer.GetTokenBalances((int)Chain.Polygon, tokenBalancesArgs);
 
                 return (balances.page, balances.balances);
             },
@@ -44,7 +44,7 @@ public class SequenceDemo : MonoBehaviour
                         new Page { page = pageNumber }
                     );
                     var balances = await Indexer.GetTokenBalances(
-                        Chain.Polygon,
+                        (int)Chain.Polygon,
                         tokenBalancesArgs
                     );
